Move tempo functions in MIDI import to separate files
